Commit Date: 11/2/2025
Some small changes to the websockets code

+++ Python Projects/ESP Websockets/constants.py
@@ -16,17 +16,17 @@ def getPos2():
     return "28"
 
 
-def get_stepPS1(x: int) -> str:
+def get_stepPS1(x: float) -> str:
     data = f"11{x}"
     return data
 
 
-def get_stepPS2(x: int) -> str:
+def get_stepPS2(x: float) -> str:
     data = f"21{x}"
     return data
 
 
-def get_global_stepsPS(x: int) -> str:
+def get_global_stepsPS(x: float) -> str:
     data = f"01{x}"
     return data
